requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai-whisper==20231117
ffmpeg-python==0.2.0
transformers==4.55.4
torch
httpx==0.27.0
pyttsx3==2.90
sentencepiece==0.2.0
accelerate==0.32.1
